pyproject.toml
[tool.poetry]
name = "ad-media"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "ad_media"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
pip = "^23.2.1"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.7"
requests = "^2.31.0"
drf-spectacular = "^0.26.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
